requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
gunicorn==23.0.0
kombu==5.4.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
flower==2.0.1